+++ sdk/resourcemanager/managedservices/armmanagedservices/ze_generated_example_registrationassignments_client_test.go
@@ -108,7 +108,7 @@ func ExampleRegistrationAssignmentsClient_BeginCreateOrUpdate() {
 }
 
 // Generated from example definition: https://github.com/Azure/azure-rest-api-specs/tree/main/specification/managedservices/resource-manager/Microsoft.ManagedServices/preview/2022-01-01-preview/examples/GetRegistrationAssignments.json
-func ExampleRegistrationAssignmentsClient_List() {
+func ExampleRegistrationAssignmentsClient_NewListPager() {
 	cred, err := azidentity.NewDefaultAzureCredential(nil)
 	if err != nil {
 		log.Fatalf("failed to obtain a credential: %v", err)
@@ -120,7 +120,7 @@ func ExampleRegistrationAssignmentsClient_List() {
 		log.Fatalf("failed to create client: %v", err)
 		return
 	}
-	pager := client.List("<scope>",
+	pager := client.NewListPager("<scope>",
 		&armmanagedservices.RegistrationAssignmentsClientListOptions{ExpandRegistrationDefinition: nil,
 			Filter: nil,
 		})